Mostly trivial
* Allow passing using null in data entity blueprint and update
* Move the Notification class out of EntityAndPendingNotifications
* Javadoc corrections
